windsurf-project/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy>=3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-Cors==4.0.0
Flask-SocketIO==5.3.6
Flask-WTF==1.2.1
Flask-Assets==2.1.0
python-dotenv==1.0.0
SQLAlchemy>=2.0.40
Pillow
python-telegram-bot==20.7
transformers
torch
geopy==2.4.0
phonenumbers==8.13.36
requests>=2.31.0
python-jose==3.3.0
python-multipart==0.0.6
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
aiohttp
python-socketio==5.10.0
redis==5.0.1
cssmin==0.2.0
jsmin==3.0.1
libsass==0.22.0
